+++ src/api/v1/users.py
@@ -6,12 +6,11 @@ from http import HTTPStatus
 from typing import Annotated
 from datetime import datetime
 
-
 from async_fastapi_jwt_auth import AuthJWT
 from fastapi.security import HTTPBearer
 from fastapi.encoders import jsonable_encoder
 from fastapi.responses import JSONResponse, Response
-from fastapi import APIRouter, Depends, Header, HTTPException
+from fastapi import APIRouter, Depends, Header, HTTPException, Path, Body
 
 from core.config import JWTSettings
 from schemas.entity import (
@@ -29,6 +28,7 @@ from schemas.entity import (
 )
 from services.user_services import get_user_service, UserService
 from services.user import UserPermissionsService, get_user_permissions_service
+from services.authorization import PermissionClaimsService, get_permission_claims_service
 
 
 MAX_SESSION_NUMBER = 5
@@ -44,19 +44,31 @@ def get_config():
 
 
 @router.post(
-    '/{user_id}/role',
+    '/{user_id}/group',
     response_model=UserInDB,
     summary='Назначение роли пользователю',
     description='Выполняет добавление новой роли для пользователя',
     response_description='Информация об обновленном пользователе'
 )
-async def add_role(
-        user_id: UUID,
-        group_assign: GroupAssign,
-        user_service: UserPermissionsService = Depends(get_user_permissions_service)
+async def add_group(
+    user_id: Annotated[UUID, Path(description='Идентификатор пользователя')],
+    group_assign: Annotated[GroupAssign, Body(description='Шаблон создания роли для пользователя')],
+    user_service: Annotated[UserPermissionsService, Depends(get_user_permissions_service)],
+    permission_claims_service: Annotated[PermissionClaimsService, Depends(get_permission_claims_service)],
+    authorize: Annotated[AuthJWT, Depends()],
+    access_token: Annotated[str, Depends(security)]
 ):
+    await authorize.jwt_required(token=access_token)
+    is_authorized = await permission_claims_service.required_permissions(
+        await authorize.get_jwt_subject(), ['add_group']
+    )
+
+    if not is_authorized:
+        raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail='Not enough rights')
+
     group_assign_encoded = jsonable_encoder(group_assign)
     user = await user_service.add_role_to_user(user_id, group_assign_encoded)
+
     if not user:
         raise HTTPException(
             status_code=HTTPStatus.NOT_FOUND,
@@ -66,17 +78,28 @@ async def add_role(
 
 
 @router.delete(
-    '/{user_id}/role',
+    '/{user_id}/group',
     response_model=UserInDB,
     summary='Создание роли',
     description='Выполняет создание новой роли',
     response_description='Информация о роли, записанной в базу данных'
 )
-async def delete_role(
-        user_id: UUID,
-        group_assign: GroupAssign,
-        user_service: UserPermissionsService = Depends(get_user_permissions_service)
+async def delete_group(
+    user_id: Annotated[UUID, Path(description='Идентификатор пользователя')],
+    group_assign: Annotated[GroupAssign, Body(description='Шаблон удаления роли для пользователя')],
+    user_service: Annotated[UserPermissionsService, Depends(get_user_permissions_service)],
+    authorize: Annotated[AuthJWT, Depends()],
+    permission_claims_service: Annotated[PermissionClaimsService, Depends(get_permission_claims_service)],
+    access_token: Annotated[str, Depends(security)]
 ):
+    await authorize.jwt_required(token=access_token)
+    is_authorized = await permission_claims_service.required_permissions(
+        await authorize.get_jwt_subject(), ['delete_group', ]
+    )
+
+    if not is_authorized:
+        raise HTTPException(status_code=HTTPStatus.FORBIDDEN, detail='Not enough rights')
+
     group_assign_encoded = jsonable_encoder(group_assign)
     user = await user_service.delete_role_from_user(user_id, group_assign_encoded)
     if not user:
